rendu/devweb/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
torch>=2.1.0
transformers>=4.45.0
peft>=0.12.0
accelerate>=0.34.0
bitsandbytes>=0.43.0
